+++ bulldozer/pipeline/bulldozer_pipeline.py
@@ -160,12 +160,9 @@ def dsm_to_dtm(config_path: str = None, **kwargs: int) -> None:
             eomanager.release(key=inner_nodata_mask_key)
 
         # Step 3: Fill the input DSM and compute the uncertainties
-        #TODO - Hotfix to remove
-        unfilled_dsm_mask_key = eomanager.create_image(eomanager.get_profile(regular_mask_key))
         fill_outputs = preprocess_dsm_filler.fill_dsm(dsm_key=input_dsm_key,
                                                       regular_key=regular_mask_key,
                                                       border_nodata_key=border_nodata_mask_key,
-                                                      unfilled_dsm_mask_key=unfilled_dsm_mask_key,
                                                       nodata=pipeline_nodata,
                                                       max_object_size=params["max_object_size"],
                                                       eomanager=eomanager,
@@ -173,11 +170,6 @@ def dsm_to_dtm(config_path: str = None, **kwargs: int) -> None:
                                                       dev_dir=developer_dir)
 
         filled_dsm_key = fill_outputs["filled_dsm"]
-
-        #TODO - Hotfix to remove
-        unfilled_dsm_mask_key = fill_outputs["unfilled_dsm_mask_key"]
-        if params["developer_mode"]:
-            eomanager.write(key=unfilled_dsm_mask_key, img_path=os.path.join(developer_dir, "unfilled_dsm_mask.tif"), binary=True)
 
         if params["developer_mode"]:
             filled_dsm_path: str = os.path.join(developer_dir, "filled_dsm.tif")
@@ -256,7 +248,7 @@ def dsm_to_dtm(config_path: str = None, **kwargs: int) -> None:
 
         # Step 7: remove pits
         BulldozerLogger.log("Pits removal: Starting.", logging.INFO)
-        dtm_key, pits_mask_key = fill_pits.run(dtm_key, border_nodata_mask_key, unfilled_dsm_mask_key, eomanager)
+        dtm_key, pits_mask_key = fill_pits.run(dtm_key, border_nodata_mask_key, eomanager)
         eomanager.write(key=pits_mask_key, img_path=os.path.join(output_masks_dir, "filled_pits.tif"), binary=True)
         BulldozerLogger.log("Pits removal: Done.", logging.INFO)
         eomanager.release(key=pits_mask_key)
@@ -266,9 +258,6 @@ def dsm_to_dtm(config_path: str = None, **kwargs: int) -> None:
         final_dtm = eomanager.get_array(key=dtm_key)[0]
         border_nodata_mask = eomanager.get_array(key=border_nodata_mask_key)[0]
         final_dtm[border_nodata_mask==1] = input_nodata
-        #TODO - Hotfix to remove
-        unfilled_dsm_mask = eomanager.get_array(key=unfilled_dsm_mask_key)[0]
-        final_dtm[unfilled_dsm_mask==1] = input_nodata
         BulldozerLogger.log("Applying border no data: Done...", logging.INFO)
 
         # Write final outputs
@@ -294,8 +283,6 @@ def dsm_to_dtm(config_path: str = None, **kwargs: int) -> None:
             # if the DHM is not generated, release the border_nodata_mask memory
             eomanager.release(key=border_nodata_mask_key)
         eomanager.release(key=filled_dsm_key)
-        #TODO - Hotfix to remove
-        eomanager.release(key=unfilled_dsm_mask_key)
 
         # write final dtm
         eomanager.write(key=dtm_key, img_path=os.path.join(params["output_dir"], "dtm.tif"))